pyproject.toml
[tool.poetry]
name = "vc-service-back"
version = "0.1.0"
description = ""
authors = ["h3nryfost3r <henry.foster71@yahoo.com>"]

[tool.poetry.dependencies]
python = "^3.10"
fastapi = "^0.103.1"
sqlalchemy = "^2.0.20"
alembic = "^1.12.0"
pydantic-settings = "^2.0.3"
asyncpg = "^0.28.0"
strawberry-graphql = "^0.209.2"
uvicorn = "^0.23.2"

[tool.poetry.group.dev.dependencies]
pytest = "^7.4.1"
ruff = "^0.0.287"
alembic = "^1.12.0"
pre-commit = "^3.4.0"
mypy = "^1.5.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
